Move support component to own folder
